src/components/problem-content/requirements.txt
pyyaml==6.0.2
supabase==2.0.3
python-dotenv
openai